e2e: move the kwok download helpers into a test file
The ap-lint unused check runs on non-test code and flagged ensureKwok in
binaries.go, which only TestCaptureKwok calls. Move it (with its stage list
and downloadTo helper) into kwok_test.go; no behaviour change.
