+++ tests/e2e/binaries.go
@@ -156,74 +156,6 @@ func ensureKubeAPIServerFromEnvtest(ctx context.Context, t *testing.T) string {
 	return binPath
 }
 
-// defaultKwokVersion is the kwok release used for traffic captures. Override
-// with KWOK_VERSION.
-const defaultKwokVersion = "v0.8.0"
-
-// kwokStages are the Stage configs (relative to kustomize/stage in the kwok
-// repo) that make the kwok controller behave like kubelets: nodes register
-// and go Ready, renew their leases and report status, and pods go Running,
-// complete and delete. The standalone binary ships without any.
-var kwokStages = []string{
-	"node/fast/node-initialize.yaml",
-	"node/heartbeat-with-lease/node-heartbeat-with-lease.yaml",
-	"pod/fast/pod-ready.yaml",
-	"pod/fast/pod-complete.yaml",
-	"pod/fast/pod-delete.yaml",
-}
-
-// kwokInstall is a kwok controller binary with its stage configs.
-type kwokInstall struct {
-	binPath string
-	stages  []string
-}
-
-// ensureKwok returns a kwok controller for the current platform, downloading
-// and caching it (and its stage configs) under .build/ if necessary.
-func ensureKwok(ctx context.Context, t *testing.T) *kwokInstall {
-	t.Helper()
-	version := os.Getenv("KWOK_VERSION")
-	if version == "" {
-		version = defaultKwokVersion
-	}
-	cacheDir := filepath.Join(repoRoot(t), ".build", "e2e-bin", "kwok-"+version)
-	if err := os.MkdirAll(filepath.Join(cacheDir, "stages"), 0o755); err != nil {
-		t.Fatalf("mkdir %s: %v", cacheDir, err)
-	}
-
-	install := &kwokInstall{binPath: filepath.Join(cacheDir, "kwok")}
-	if fi, err := os.Stat(install.binPath); err != nil || !fi.Mode().IsRegular() {
-		url := fmt.Sprintf("https://github.com/kubernetes-sigs/kwok/releases/download/%s/kwok-%s-%s", version, runtime.GOOS, runtime.GOARCH)
-		t.Logf("downloading kwok %s from %s", version, url)
-		downloadTo(ctx, t, url, install.binPath, 0o755)
-	}
-	for _, stage := range kwokStages {
-		path := filepath.Join(cacheDir, "stages", filepath.Base(stage))
-		if fi, err := os.Stat(path); err != nil || !fi.Mode().IsRegular() {
-			url := fmt.Sprintf("https://raw.githubusercontent.com/kubernetes-sigs/kwok/%s/kustomize/stage/%s", version, stage)
-			downloadTo(ctx, t, url, path, 0o644)
-		}
-		install.stages = append(install.stages, path)
-	}
-	return install
-}
-
-// downloadTo downloads url to dest atomically with the given mode.
-func downloadTo(ctx context.Context, t *testing.T, url, dest string, mode os.FileMode) {
-	t.Helper()
-	tmp := dest + ".tmp"
-	if err := download(ctx, url, tmp); err != nil {
-		_ = os.Remove(tmp)
-		t.Fatalf("download %s: %v", url, err)
-	}
-	if err := os.Chmod(tmp, mode); err != nil {
-		t.Fatalf("chmod %s: %v", tmp, err)
-	}
-	if err := os.Rename(tmp, dest); err != nil {
-		t.Fatalf("rename %s into place: %v", dest, err)
-	}
-}
-
 func setupEnvtestRef() string {
 	if v := os.Getenv("SETUP_ENVTEST_VERSION"); v != "" {
 		return v
